rename "Navigator" package to "inventory"
