+++ src/main/java/de/tdrstudios/lobbyplugin/inventory/Cosmetics.java
@@ -1,4 +1,4 @@
-package de.tdrstudios.lobbyplugin.Navigator;
+package de.tdrstudios.lobbyplugin.inventory;
 
 import de.tdrstudios.lobbyplugin.LobbyPlugin;
 import org.bukkit.Bukkit;
